requirements.txt
amqp==2.5.2
astroid==2.3.3
autopep8==1.4.4
billiard==3.6.1.0
celery==4.3.0
dj-database-url==0.5.0
Django==2.2.7
django-environ==0.4.5
django-heroku==0.3.1
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
gunicorn==20.0.4
importlib-metadata==1.0.0
isort==4.3.21
kombu==4.6.6
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.0.0
psycopg2==2.8.4
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.4.4
pytz==2019.3
redis==3.3.11
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
vine==1.3.0
whitenoise==4.1.4
wrapt==1.11.2
zipp==0.6.0
